KAFKA-3333: Refactors RRP out of internals into producer
